+++ src/bas_metadata_library/standards/iso_19115_1/__init__.py
@@ -93,10 +93,10 @@ class MetadataRecordConfigV2(_MetadataRecordConfig):
 
     def dump(self, file: Path) -> None:
         with file.open(mode="w") as file:
-            json.dump(encode_config_for_json(config=deepcopy(self.config)), file)
+            json.dump(encode_config_for_json(config=deepcopy(self.config)), file, indent=2)
 
     def dumps(self) -> str:
-        return json.dumps(encode_config_for_json(config=deepcopy(self.config)))
+        return json.dumps(encode_config_for_json(config=deepcopy(self.config)), indent=2)
 
 
 class MetadataRecordConfigV3(_MetadataRecordConfig):